load .env and serverInjection as part of config
